% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abundExamples}
\alias{abundExamples}
\title{Example relative abundance estimates from simulated data}
\format{
A list with 100 mcmc (coda) estimates in it.
}
\usage{
abundExamples
}
\description{
A dataset containing mcmc relative abundance estimates from simulated
BBS-type data from Cohen et al. (2018). Each estimate can
be used in \code{estMC} function to estimate MC with uncertainty.
}
\keyword{datasets}
