% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConnectivity.R
\name{calcMantel}
\alias{calcMantel}
\title{Calculate Mantel correlation (rM) from points.}
\usage{
calcMantel(targetPoints = NULL, originPoints = NULL,
  targetDist = NULL, originDist = NULL)
}
\arguments{
\item{targetPoints}{A \code{SpatialPoints} object, with length number of
animals tracked.  Each point indicates the point estimate location in
the non-release season.}

\item{originPoints}{A \code{SpatialPoints} object, with length number of
animals tracked.  Each point indicates the release location of an animal.}

\item{fasterDist}{If TRUE, calculates distances using a sphere approximation
of the earth; if FALSE, uses the more accurate (but slower) ellipsoid
approximation. Defaults to FALSE, which is generally fast enough.}
}
\value{
\code{calcMantel} returns the Mantel correlation, calculated without
   accounting for any uncertainty.
}
\description{
Calculation of rM from SpatialPoints geolocators and/or GPS
data, not accounting for uncertainty.
}
\examples{
rM0 <- calcMantel(originPoints = OVENdata$originPoints, # Capture Locations
                  targetPoints = OVENdata$targetPoints) # Target locations
str(rM0)
}
\references{
Cohen, E. B., J. A. Hostetler, M. T. Hallworth, C. S. Rushing, T. S. Sillett,
and P. P. Marra. 2018. Quantifying the strength of migratory connectivity.
Methods in Ecology and Evolution 9: 513 - 524.
\href{http://doi.org/10.1111/2041-210X.12916}{doi:10.1111/2041-210X.12916}
}
\seealso{
\code{\link{estMantel}}
}
