% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcConnectivity.R
\name{calcMC}
\alias{calcMC}
\title{Migratory connectivity strength function}
\usage{
calcMC(originDist, targetDist, psi, originRelN)
}
\arguments{
\item{originDist}{}

\item{targetDist}{}

\item{psi}{}

\item{originRelN}{}
}
\value{
real value between -1 and 1, inclusive (usually between 0 and 1)
   indicating the degree of migratory connectivity.
}
\description{
Migratory connectivity strength function
}

