% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcConnectivity.R
\name{calcMCSmall}
\alias{calcMCSmall}
\title{Migratory connectivity strength function corrected for small absolute
abundances}
\usage{
calcMCSmall(originDist, targetDist, psi, originAbund)
}
\arguments{
\item{originDist}{Distances between the B origin sites.  Symmetric B by B
matrix.}

\item{targetDist}{Distances between the W target sites.  Symmetric W by W
matrix.}

\item{psi}{Transition probabilities between B origin and W target sites.
Matrix with B rows and W columns where rows sum to 1.}

\item{originAbund}{Actual or estimated abundances at B origin sites.
   Numeric vector of length B with all values greater than or equal to 1.
   These should be on the scale of number of individuals.}
}
\value{
real value between -1 and 1, inclusive (usually between 0 and 1)
   indicating the degree of migratory connectivity.
}
\description{
Migratory connectivity strength function corrected for small absolute
abundances
}
\seealso{
calcMC
}

