% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmrExamples}
\alias{cmrExamples}
\title{Example RMark transition probability estimates from simulated data}
\format{A list with 100 RMark estimates in it.}
\usage{
cmrExamples
}
\description{
A dataset containing RMark transition probability estimates from simulated
mark-recapture-recovery data from Cohen et al. (2014).  These all represent
the intermediate scenario for all settings (moderate connectivity, low
re-encounter, 100,000 banded in each breeding area).  Each estimate can
be used in \code{estMC} function to estimate MC with uncertainty.
}
\keyword{datasets}

