% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estConnectivity.R
\name{estMCGlGps}
\alias{estMCGl}
\alias{estMCGlGps}
\alias{estMCGps}
\title{Resampling of uncertainty from geolocators and/or GPS data}
\usage{
estMCGlGps(isGL, geoBias, geoVCov, originRelAbund, originDist, targetDist,
  targetPoints, targetSites, targetAssignment = NULL, originPoints = NULL,
  originSites = NULL, originAssignment = NULL, originNames = NULL,
  targetNames = NULL, nBoot = 1000, verbose = 0, nSim = 1000,
  calcCorr = T, alpha = 0.05)

estMCGl(geoBias, geoVCov, originRelAbund, originDist, targetDist, targetPoints,
  targetSites, targetAssignment = NULL, originPoints = NULL,
  originSites = NULL, originAssignment = NULL, originNames = NULL,
  targetNames = NULL, nBoot = 1000, verbose = 0, nSim = 1000,
  calcCorr = T, alpha = 0.05)

estMCGps(originPoints, targetPoints, originSites, targetSites, originRelAbund,
  originDist, targetDist, originNames = NULL, targetNames = NULL,
  nBoot = 1000, verbose = 0, calcCorr = T, alpha = 0.05)
}
\arguments{
\item{isGL}{Vector indicating which animals were tracked with geolocators.
Should be length of number of animals tracked, with TRUE for animals in
\code{targetPoints} with geolocators and FALSE for animals with GPS.}

\item{geoBias}{Vector of length 2 indicating expected bias in longitude and
latitude of \code{targetPoints}, in meters.}

\item{geoVCov}{2x2 matrix with expected variance/covariance in longitude and
latitude of \code{targetPoints}, in meters.}

\item{originRelAbund}{Relative abundances at B origin sites. A numeric
vector of length B that sums to 1.}

\item{originDist}{Distances between the B origin sites.  Symmetric B by B
matrix.}

\item{targetDist}{Distances between the W target sites.  Symmetric W by W
matrix.}

\item{targetPoints}{A \code{SpatialPoints} object, with length number of
animals tracked.  Each point indicates the point estimate location in
the non-release season.}

\item{targetSites}{Geographic definition of sites in the non-release season.}

\item{targetAssignment}{Optional. Point estimate assignment of
\code{targetPoints} to non-release season sites. Integer vector with
length number of animals tracked.}

\item{originPoints}{A \code{SpatialPoints} object, with length number of
animals tracked.  Each point indicates the release location of an animal.}

\item{originSites}{Geographic definition of sites in the release season.}

\item{originAssignment}{Assignment of \code{originPoints} to release season
sites. Integer vector with length number of animals tracked. Optional,
but either \code{originAssignment} or \code{originSites} and
\code{originPoints} should be defined.}

\item{originNames}{Optional. Vector of names for the release season sites.}

\item{targetNames}{Optional. Vector of names for the non-release season
sites.}

\item{nBoot}{Number of bootstrap runs. Animals are sampled with replacement
for each of these to estimate sampling uncertainty.}

\item{verbose}{Integer 0-3 for level of output during bootstrap: 0 = none,
1 = every 10, 2 = every run, 3 = every animal.}

\item{nSim}{Number of times to sample random points for each animal from
parametric distribution of non-release season error. Ignored for GPS
points (assumed to have no geographic error).}

\item{calcCorr}{In addition to MC, should function also estimate continuous
correlation between release and non-release locations?  Default is TRUE.}

\item{alpha}{Level for confidence intervals provided.}
}
\value{
A list with elements:
\describe{
  \item{\code{sampleMC}}{\code{nBoot} sampled values for MC. Provided
     to allow the user to compute own summary statistics.}
  \item{\code{samplePsi}}{Array of sampled values for psi. \code{nBoot} x
     [number of origin sites] x [number of target sites]. Provided
     to allow the user to compute own summary statistics.}
  \item{\code{pointSites}}{Matrix of point assignment of number of animals
     to each origin and target site combination.}
  \item{\code{pointPsi}}{Simple point estimate of psi matrix, as
     \code{prop.table(pointSites, 1)}.}
  \item{\code{pointMC}}{Simple point estimate of MC, using \code{pointPsi}
     and \code{originRelAbund}.}
  \item{\code{meanMC, medianMC}}{Mean and median of \code{sampleMC}.
     Estimates of MC incorporating parametric uncertainty.}
  \item{\code{seMC}}{Standard error of MC, estimated from SD of
     \code{sampleMC}.}
  \item{\code{simpleCI}}{\code{1 - alpha} confidence interval for MC,
     estimated as \code{alpha/2} and \code{1 - alpha/2} quantiles of
     \code{sampleMC}.}
  \item{\code{bcCI}}{Bias-corrected \code{1 - alpha} confidence interval
     for MC.  Preferable to \code{simpleCI} when \code{pointMC} is the
     best estimate of MC. \code{simpleCI} is preferred when
     \code{medianMC} is a better estimator. When \code{pointMC==medianMC},
     these should be equivalent.}
  \item{\code{sampleCorr}}{\code{nBoot} sampled values for continuous
     correlation. Provided to allow the user to compute own summary
     statistics.}
  \item{\code{pointCorr}}{Simple point estimate of continuous correlation,
     using \code{originPoints} and \code{targetPoints}.}
  \item{\code{meanCorr, medianCorr, seCorr, simpleCICorr, bcCICorr}}{Summary
     statistics for continuous correlation bootstraps.}
}
}
\description{
Resampling of uncertainty from geolocators and/or GPS data
}
\section{Functions}{
\itemize{
\item \code{estMCGl}: Convenience function if all points are from geolocators.

\item \code{estMCGps}: Convenience function if all points are from GPS.
}}

