% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estConnectivity.R
\name{estStrength}
\alias{estStrength}
\title{Estimate MC, migratory connectivity strength}
\usage{
estStrength(
  originDist,
  targetDist,
  originRelAbund,
  psi,
  sampleSize = NULL,
  originSites = NULL,
  targetSites = NULL,
  originNames = NULL,
  targetNames = NULL,
  nSamples = 1000,
  row0 = 0,
  verbose = 0,
  alpha = 0.05,
  approxSigTest = FALSE,
  sigConst = 0,
  maintainLegacyOutput = FALSE
)
}
\arguments{
\item{originDist}{Distances between the B origin sites. Symmetric B by B
matrix}

\item{targetDist}{Distances between the W target sites. Symmetric W by W
matrix}

\item{originRelAbund}{Relative abundance estimates at B origin sites. Either
a numeric vector of length B that sums to 1 or an mcmc object with at least
\code{nSamples} rows and columns including 'relN[1]' through 'relN[B]'}

\item{psi}{Transition probabilities between B origin and W target sites.
Either a matrix with B rows and W columns where rows sum to 1, an array with
dimensions x, B, and W (with x samples of the transition probability matrix
from another model), an 'estPsi' object (result of calling estTransition),
or a MARK object with estimates of transition probabilities}

\item{sampleSize}{Total sample size of animals that psi will be estimated
from. Should be the number of animals released in one of the origin sites
and observed in one of the target sites (or vice-versa). Optional, but
recommended, unless psi is an estPsi object (in which case this function can
pull it from there)}

\item{originSites}{If \code{psi} is a MARK object, this must be a numeric
vector indicating which sites are origin}

\item{targetSites}{If \code{psi} is a MARK object, this must be a numeric
vector indicating which sites are target}

\item{originNames}{Optional. Vector of names for the origin sites}

\item{targetNames}{Optional. Vector of names for the target sites}

\item{nSamples}{Number of times to resample \code{psi} and/or
\code{originRelAbund}. The purpose is to estimate sampling uncertainty;
higher values here will do so with more precision}

\item{row0}{If \code{originRelAbund} is an mcmc object, this can be set
to 0 (default) or any greater integer to specify where to stop ignoring
samples ("burn-in")}

\item{verbose}{0 (default) to 2. 0 prints no output during run. 1 prints
a progress update and summary every 100 samples. 2 prints a
progress update and summary every sample}

\item{alpha}{Level for confidence/credible intervals provided. Default (0.05)
gives 95 percent CI}

\item{approxSigTest}{Should function compute approximate one-sided
significance tests (p-values) for MC from the resampling? Default is
FALSE}

\item{sigConst}{Value to compare MC to in significance test. Default is 0}

\item{maintainLegacyOutput}{version 0.4.0 of \code{MigConnectivity}
updated the structure of the estimates. If you have legacy code that refers
to elements within an \code{estMigConnectivity} object (results of
\code{estMC}), you can set this to TRUE to also keep the old structure.
Defaults to FALSE}
}
\value{
\code{estMC} returns a list with the elements:
\describe{
  \item{\code{MC}}{List containing estimates of migratory connectivity
   strength:
   \itemize{
     \item{\code{sample}} \code{nSamples} sampled values for
      MC. Provided to allow the user to compute own summary statistics.
     \item{\code{mean}} Mean of \code{MC$sample}. Main estimate of MC,
      incorporating parametric uncertainty.
     \item{\code{se}} Standard error of MC, estimated from SD of
      \code{MC$sample}.
     \item{\code{simpleCI}} Default\code{1 - alpha} confidence interval for
      MC, estimated as \code{alpha/2} and \code{1 - alpha/2} quantiles of
      \code{MC$sample}.
     \item{\code{bcCI}} Bias-corrected \code{1 - alpha} confidence interval
      for MC. May be preferable to \code{MC$simpleCI} when \code{MC$mean} is
      the best estimate of MC. \code{MC$simpleCI} is preferred when
      \code{MC$median} is a better estimator. When \code{MC$mean==MC$median},
      these should be identical.  Estimated as the
      \code{pnorm(2 * z0 + qnorm(alpha / 2))} and
      \code{pnorm(2 * z0 + qnorm(1 - alpha / 2))} quantiles of \code{MC$sample},
      where z0 is the proportion of \code{MC$sample < MC$mean}.
     \item{\code{hpdCI}} \code{1 - alpha} credible interval for MC,
      estimated using the highest posterior density (HPD) method.
     \item{\code{median}} Median of MC, alternate point estimate also
      including parametric uncertainty.
     \item{\code{point}} Simple point estimate of MC, using the point
     estimates of \code{psi} and \code{originRelAbund} (usually the mean
     values), not accounting for sampling error.
     \item{\code{simpleP}} Approximate p-value for MC, estimated as the
     proportion of bootstrap iterations where MC < \code{sigConst} (or MC >
     \code{sigConst} if \code{pointMC < sigConst}).  Note that if the
     proportion is 0, a default value of 0.5 / \code{nSamples} is provided,
     but this is best interpreted as p < 1 / \code{nSamples}.  NULL when
     \code{approxSigTest==FALSE}.
     \item{\code{bcP}} Approximate bias-corrected p-value for MC, estimated as
     \code{pnorm(qnorm(simpleP) - 2 * z0)}, where z0 is the proportion of
     \code{sampleMC < meanMC}.  May be a better approximation of the p-value
     than \code{simpleP}, but many of the same limitations apply.  NULL when
     \code{approxSigTest==FALSE}.
   }
  }
  \item{\code{input}}{List containing the inputs to \code{estStrength}.}
}
}
\description{
Resampling of uncertainty for MC (migratory connectivity strength)
from estimates of psi (transition probabilities) and/or relative abundance.
Psi estimates can come from RMark psi matrix, MCMC samples, or other samples
expressed in array form. Abundance estimates for each origin site and be
either just point estimates (no uncertainty propagated) or MCMC samples.
Other inputs include distances between origin sites, distances between target
sites, and sample size used to estimate psi.
}
\seealso{
\code{\link{calcMC}}, \code{\link{estTransition}},
  \code{\link{estMC}}, \code{\link{estMantel}},
  \code{\link{plot.estMigConnectivity}}
}
