% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo.bias}
\alias{geo.bias}
\title{Archival light-level geolocator bias estimates}
\format{A vector of length two with bias esimates in geolocator locations.}
\usage{
geo.bias
}
\description{
Location bias estimates in light-level geolocator estimates calculated
using birds captured at known locations in Florida, Jamaica and Puerto Rico.
Location bias is reported in meters.
}
\keyword{datasets}

