% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo.vcov}
\alias{geo.vcov}
\title{Archival light-level geolocator covariance matrix}
\format{A 2x2 matrix of covariance estimates.}
\usage{
geo.vcov
}
\description{
Covariance estimates in light-level geolocator estimates calculated
using birds captured at known locations in Florida, Jamaica and Puerto Rico.
Covariance is reported in meters.
}
\keyword{datasets}

