% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{getCMRexample}
\alias{getCMRexample}
\title{Grab (from GitHub server) example RMark transition probability estimates
obtained from simulated data}
\usage{
getCMRexample(number = 1)
}
\arguments{
\item{number}{Integer 1 - 100, which simulation and RMark estimate you want}
}
\value{
RMark object
}
\description{
Get a dataset containing RMark transition probability estimates from
simulated mark-recapture-recovery data from Cohen et al. (2014).  These all
represent the intermediate scenario for all settings (moderate connectivity,
low re-encounter, 100,000 banded in each breeding area).  Each estimate can
be used in \code{estMC} function to estimate MC with uncertainty.
}

