% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isGL}
\alias{isGL}
\title{Archival light-level geolocator or PinPoint-10 GPS tag}
\format{\code{logical} of length 39}
\usage{
isGL
}
\description{
\code{logical} vector indicating whether location estimates were obtained with a
light-level geolocator (\code{TRUE}) or PinPoint-10 GPS tag (\code{FALSE})
}
\keyword{datasets}

