% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConnectivity.R
\name{optimSites}
\alias{optimSites}
\title{Optimal site arrangement on origin and/or target side}
\usage{
optimSites(
  originSamples,
  targetSamples,
  originBlocks,
  targetBlocks,
  algorithm = "maxMC",
  originRelAbund = NULL,
  originFixed = FALSE,
  targetFixed = FALSE,
  originTies = NULL,
  targetTies = NULL,
  verbose = 0
)
}
\arguments{
\item{originSamples}{which blocks got pulled}

\item{targetSamples}{which blocks got pulled}

\item{originBlocks}{origin range divided up into blocks around telemetry and
geolocator points}

\item{targetBlocks}{target range divided up into blocks around telemetry and
geolocator points}

\item{algorithm}{"maxMC" or something that hasn't been set up yet}

\item{originRelAbund}{for calculating MC}

\item{originFixed}{if TRUE, only optimizes target side}

\item{targetFixed}{if TRUE, only optimizes origin side}

\item{originTies}{not used yet}

\item{targetTies}{not used yet}
}
\value{
list
}
\description{
Optimal site arrangement on origin and/or target side
}
