% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{originDist}
\alias{originDist}
\title{Distance between Ovenbird breeding distribution origin sites}
\format{square distance matrix}
\usage{
originDist
}
\description{
The pairwise Great Circle Distance between the center of the polygons contained within
\code{originSites}. See "Ovenbird breeding distribution origin sites" or \code{originSites}
}
\keyword{datasets}

