% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{originPoints}
\alias{originPoints}
\title{Ovenbird capture locations}
\format{SpatialPoints in Lambert Conformal Conic}
\usage{
originPoints
}
\description{
Capture locations for 39 Ovenbirds caught during the breeding season who carried
either a light-level geolocator or PinPoint-10 GPS tag. Ovenbirds were captured
at Hubbard Brook Experimental Forest, NH and Jug Bay Wetland Sanctuary, MD. These
data are used as \code{originPoints} in the \code{estMC} function. \code{coords.x1}
and \code{coords.x2} represent the longitude and latitude of the capture sites,
respectively. The data are projected in Lambert Conformal Conic.
}
\details{
"+proj=aea +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80
+datum=NAD83 +units=m +no_defs +towgs84=0,0,0"
}
\keyword{datasets}

