% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{originRelAbund}
\alias{originRelAbund}
\title{Relative abundance estimates of Ovenbird within two breeding regions}
\format{A vector of length two with relative abundance estimates.}
\usage{
originRelAbund
}
\description{
A dataset containing relative abundance estimates from BBS data
reported in Cohen et al. (in prep). These estimates can
be used in \code{estMC} function as \code{originRelAbund} in conjuction
with archival light-level geolocator and GPS locations.
}
\keyword{datasets}

