% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{originSites}
\alias{originSites}
\title{Ovenbird breeding distribution origin sites}
\format{\code{SpatialPolygons}}
\usage{
originSites
}
\description{
Breeding distribution prigin sites used in Cohen et al. (in prep) to estimate
MC of Ovenbirds tracked with light-level geolocators and PinPoint-10 GPS tags. There
are two breeding origin sites, one that encompasses NH and another that encompasses MD capture
deployment locations.
}
\keyword{datasets}

