% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputConnectivity.R
\name{plot.intrinsicAssign}
\alias{plot.intrinsicAssign}
\title{basic plot function for the different isoAssign outputs}
\usage{
plot.intrinsicAssign(x, map, ...)
}
\description{
basic plot function for the different isoAssign outputs
}
