% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputConnectivity.R
\name{plot.isoAssign}
\alias{plot.isoAssign}
\title{basic plot function for the different isoAssign outputs}
\usage{
\method{plot}{isoAssign}(x, map, ...)
}
\description{
basic plot function for the different isoAssign outputs
}
