% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projections}
\alias{projections}
\title{Map projections}
\format{A named list of strings.}
\usage{
projections
}
\description{
Map projections used when sampling from geolocator bias/error, for
example. The argument \code{resampleProjection} in \code{estMC} and
\code{estMantel} need units = m, which is true of all of these except
WGS84 (the second). First item is Equidistant Conic, which preserves
distances around latitude = 0 and longitude = 0. This is a good general
purpose projection, but the ideal projection may depend on the locations of
your points.  See names in list for suggestions. Other potential projections
can be found at \url{http://www.spatialreference.org/ref/}
}
\keyword{datasets}
