% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleOriginN}
\alias{sampleOriginN}
\alias{sampleOriginRelN}
\title{Example origin site abundances and relative abundances}
\format{Each dataset is a named list with 5 vectors in it. Each vector has
4 elements (for the 4 origin sites).  The relative abundance vectors each
sum to 1. The 5 scenarios are:
\itemize{
  \item Base: Equal abundance at each origin site
  \item B Doub: The second origin site has twice the abundance of the other
     three sites
  \item B Half: The second origin site has half the abundance of the other
     three sites
  \item D Doub: The last origin site has twice the abundance of the other
     three sites
  \item D Half: The last origin site has half the abundance of the other
     three sites
}}
\usage{
sampleOriginN

sampleOriginRelN
}
\description{
\code{sampleOriginN} is a dataset containing example origin site abundances
from 5 scenarios used in Cohen et al. (in prep.).  For the same 5
scenarios, \code{sampleOriginRelN} contains the relative abundances.
}
\keyword{datasets}

