% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simConnectivity.R
\name{simCMRData}
\alias{simCMRData}
\title{Simulate capture-mark-reencounter (CMR) migratory movement data}
\usage{
simCMRData(psi, banded, r)
}
\arguments{
\item{psi}{Transition probabilities between B origin sites and W target
sites. B by W matrix}

\item{banded}{A vector of the number of released animals from each origin
site (including those never reencountered in a target site). Length B}

\item{r}{A vector (length W) of reencounter probabilities at each target site}
}
\value{
\code{simCMRData} returns a list with the elements:
\describe{
  \item{\code{reencountered}}{B by W matrix with numbers reencountered at
   each target site, by origin site}
  \item{\code{migrated}}{B by W matrix with numbers migrated to
   each target site, by origin site. Assumes survival to arrival is 1}
  \item{\code{input}}{List containing the inputs to function}
}
}
\description{
Simulate capture-mark-reencounter (CMR) migratory movement data
}
