% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simConnectivity.R
\name{simGeneticData}
\alias{simGeneticData}
\title{Simulate animal genoscape data for estimating migratory connectivity}
\usage{
simGeneticData(
  genPops,
  psi,
  originRelAbund,
  sampleSize,
  originSites = NULL,
  targetSites = NULL,
  captured = "target",
  requireEveryOrigin = FALSE,
  resampleNAs = TRUE,
  verbose = 0
)
}
\arguments{
\item{genPops}{Output of function \code{\link{simGeneticPops}}}

\item{psi}{Transition probabilities between B origin sites and W target
sites. B by W matrix}

\item{originRelAbund}{Vector of relative abundances at B origin sites}

\item{sampleSize}{Either the total number of data points to simulate or a
vector with the number at each target or origin site. If only the total is
provided, sampling will be done in proportion to abundance}

\item{originSites}{A polygon spatial layer (sf - MULTIPOLYGON) defining the
geographic representation of sites in the origin season}

\item{targetSites}{A polygon spatial layer (sf - MULTIPOLYGON) defining the
geographic representation of sites in the target season}

\item{captured}{Either "target" (the default) or "origin", indicating which
side animal data were collected on}

\item{requireEveryOrigin}{If TRUE, the function will throw an error if it
looks like at least one origin site has no animals released in or
migrating to it, or if it can, keep simulating until representation is
met. This helps estTransition or estMC not throw an error. Default FALSE.}

\item{resampleNAs}{if TRUE (default) re-samples locations until no NAs
found in the genetic probabilities. If \code{FALSE} equal probability is
assigned to all regions.}

\item{verbose}{- if 0 (default) no output to screen is generated. If > 0,
gives updates on stage of process.}
}
\value{
\code{simGeneticData} returns a list with the elements:
\describe{
  \item{\code{originAssignment}}{Vector with true origin site of each animal}
  \item{\code{targetAssignment}}{Vector with true target site of each animal}
  \item{\code{originPointsTrue}}{True origin location of each animal, type sf,
   same projection as originSites}
  \item{\code{targetPointsTrue}}{True target location of each animal, type sf,
   same projection as targetSites}
  \item{\code{genProbs}}{Table of assignment site probabilities for each
   animal}
  \item{\code{genRasters}}{raster stack?}
  \item{\code{input}}{List containing the inputs to function}
}
}
\description{
Simulate animal genoscape data for estimating migratory connectivity
}
