% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simConnectivity.R
\name{simGeneticPops}
\alias{simGeneticPops}
\title{Simulates genetic population rasters from input polygons}
\usage{
simGeneticPops(
  popBoundaries,
  npts = 1000,
  res = c(0.0833, 0.0833),
  bufferRegions = FALSE,
  bufferDist = NULL,
  popNames = NULL
)
}
\arguments{
\item{popBoundaries}{- a list containing polygons that represent the
boundaries of each region}

\item{npts}{- number of random points used to generate a kernel density
surface within each region}

\item{res}{- the desired resolution of the output raster}

\item{bufferRegions}{boolean if \code{TRUE} a buffer is applied to each
region to avoid or increase overlap}

\item{bufferDist}{- Desired buffer distance in meters. Positive values
enlarge regions, negative buffers make regions smaller}

\item{popNames}{- a vector the same length as popBoundaries}
}
\value{
returns a rasterStack probability surface.
}
\description{
This code simulates rasters analogous to the genetic surfaces created from
the bird genoscape project
}
