% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simConnectivity.R
\name{simGeneticPops_Overlap}
\alias{simGeneticPops_Overlap}
\title{Simulates genetic population rasters from input polygons}
\usage{
simGeneticPops_Overlap(
  bufferedBoundaries = NULL,
  maskBoundaries = NULL,
  popBoundaries,
  npts = 1000,
  res = NULL,
  bufferRegions = FALSE,
  bufferDist = -50000,
  popNames = NULL,
  verbose = 0
)
}
\arguments{
\item{bufferedBoundaries}{= buffers applied to populations to increase
assignment uncertainty. Users need to remove potential overlapping regions
where assignment is not possible}

\item{maskBoundaries}{= list of polygons that confines genetic uncertainty.
length of list = length of bufferedBoundaries/popBoundaries. The same
polygon could be used for multiple 'populations' to ensure overlap}

\item{popBoundaries}{= population boundaries - no buffers applied.}

\item{npts}{- number of random points used to generate a kernel density
surface within each region}

\item{res}{- the desired resolution of the output raster}

\item{bufferRegions}{boolean if \code{TRUE} a buffer is applied to each
region to avoid or increase overlap.}

\item{bufferDist}{- Desired buffer distance in meters. Positive values
enlarge regions, negative buffers make regions smaller. Only used when
\code{bufferRegions=TRUE}.}

\item{popNames}{- a vector the same length as popBoundaries}

\item{verbose}{- if 0 (default) no output to screen is generated. If > 0,
gives updates on stage of process.}
}
\value{
returns a rasterStack probability surface.
}
\description{
This code simulates rasters analogous to the genetic surfaces created from
the bird genoscape project
}
