% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simConnectivity.R
\name{simGL}
\alias{simGL}
\title{Simulate geolocator data}
\usage{
simGL(
  psi,
  originRelAbund,
  sampleSize,
  originSites = NULL,
  targetSites = NULL,
  captured = "origin",
  geoBias = NULL,
  geoVCov = NULL,
  geoBiasOrigin = NULL,
  geoVCovOrigin = NULL,
  S = 1,
  p = list(1, 1)
)
}
\arguments{
\item{geoVCovOrigin}{}
}
\value{

}
\description{
Simulate geolocator data
}
