% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simConnectivity.R
\name{simMove}
\alias{simMove}
\title{Simulates position of birds by individual, season, year, and month.}
\usage{
simMove(breedingAbund, breedingDist, winteringDist, psi, nYears = 10,
  nMonths = 3, winMoveRate = 0, sumMoveRate = 0, winDispRate = 0,
  sumDispRate = 0, natalDispRate = 0, breedDispRate = 0, verbose = 0)
}
\arguments{
\item{breedingAbund}{Vector with number of birds to simulate starting at
each breeding site.}

\item{breedingDist}{Distances between the breeding sites. Symmetric matrix.}

\item{winteringDist}{Distances between the wintering sites. Symmetric
matrix.}

\item{psi}{Transition probabilities between B origin and W target sites.
A matrix with B rows and W columns where rows sum to 1.}

\item{nYears}{Number of years to simulate movement.}

\item{nMonths}{Number of months per breeding and wintering season.}

\item{winMoveRate}{Within winter movement rate.  Defaults to 0 (no
movement).}

\item{sumMoveRate}{Within summer movement rate.  Defaults to 0 (no
movement).}

\item{winDispRate}{Between winter dispersal rate.  Defaults to 0 (no
dispersal).}

\item{sumDispRate}{Between summer dispersal rate.  Defaults to 0 (no
dispersal).  Setting this to a value above 0 is equivalent to setting
both natal and breeding dispersal to that same value.}

\item{natalDispRate}{Natal dispersal rate. Controls the movement of
animals from their birthplace on their first return to the breeding
grounds.  Defaults to 0 (return to the birthplace for all).}

\item{breedDispRate}{Breeding dispersal rate. Controls the movement of
animals between breeding sites on spring migrations after the first.
Defaults to 0 (return to the same breeding site each year).}

\item{verbose}{If set to a value > 0, informs the user on the passage
   of years and seasons during the simulation. Defaults to 0 (no output
   during simulation).}
}
\value{
\code{simMove} returns a list with elements:
\describe{
  \item{\code{animalLoc}}{\code{sum(breedingAbund)} (number of animals)
     by 2 by \code{nYears} by \code{nMonths} array with the simulated
     locations of each animal in each month of each season (summer or
     winter) of each year.  Values of cells are 1...B (first column) and
     1...W (second column) where B is the number of breeding sites and W is
     the number of wintering sites.}
  \item{\code{breedDispMat}}{B by B matrix of probabilities of breeding
     dispersal between each pair of 1...B breeding sites.  Direction is from
     row to column, so each row sums to 1.}
  \item{\code{natalDispMat}}{B by B matrix of probabilities of natal
     dispersal between each pair of 1...B breeding sites.  Direction is from
     row to column, so each row sums to 1.}
  \item{\code{sumMoveMat}}{B by B matrix of probabilities of within season
     movement between each pair of 1...B breeding sites.  Direction is from
     row to column, so each row sums to 1.}
  \item{\code{winDispMat}}{W by W matrix of probabilities of dispersal
     between each pair of 1...W nonbreeding sites.  Direction is from
     row to column, so each row sums to 1.}
  \item{\code{winMoveMat}}{W by W matrix of probabilities of within season
     movement between each pair of 1...W nonbreeding sites.  Direction is
     from row to column, so each row sums to 1.}
}
}
\description{
Incorporates migratory connectivity, movement within season, and dispersal
between seasons. Does not incorporate births or deaths.
}

