% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetDist}
\alias{targetDist}
\title{Distance between Ovenbird non-breeding distribution target sites}
\format{square distance matrix}
\usage{
targetDist
}
\description{
The pairwise Great Circle Distance between the center of the polygons contained within
\code{targetSites}. See "Ovenbird non-breeding distribution target sites" or \code{targetSites}
}
\keyword{datasets}

