% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{targetSites}
\alias{targetSites}
\title{Ovenbird non-breeding distribution target sites}
\format{\code{SpatialPolygons}}
\usage{
targetSites
}
\description{
Non-breeding distribution target sites used in Cohen et al. (in prep) to estimate
MC of Ovenbirds tracked with light-level geolocators and PinPoint-10 GPS tags. There
are three non-breeding target sites 1) Florida, United States, 2) Cuba, and 3) Hispaniola
(Dominican Republic and Haiti).
}
\keyword{datasets}

