% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConnectivity.R
\name{calcNMC}
\alias{calcNMC}
\title{Calculate NMC_XY, another type of migratory connectivity strength}
\usage{
calcNMC(psi)
}
\arguments{
\item{psi}{Matrix of transition probabilities}
}
\value{
\code{calcNMC} returns a list with elements:
\describe{
  \item{\code{NMC}}{scalar real value between 0 and 1, indicating the
   strength of network migratory connectivity}
  \item{\code{NMCpop}}{Vector of network migratory connectivity diversity
   values, also between 0 and 1, the X-node-specific version of NMC_XY}
}
}
\description{
Provides simple calculation of NMC_XY (network migratory connectivity
strength between seasons X and Y) and network migratory connectivity
diversity (X node-specific version of NMC_XY) from point estimate of psi
(transition probabilities). Does not include measures of uncertainty.
}
\examples{
nScenarios <- length(samplePsis)
NMC1 <- vector("list", nScenarios)
for (i in 1:nScenarios) {
  NMC1[[i]] <- calcNMC(samplePsis[[i]])
}
names(NMC1) <- names(samplePsis)
str(NMC1)
}
\seealso{
\code{\link{estNMC}}, \code{\link{calcMC}}, \code{\link{estMC}}
}
