% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estConnectivity.R
\name{diffMC}
\alias{diffMC}
\title{Pairwise differences between two or more independent MC estimates}
\usage{
diffMC(estimates, nSamples = NULL, alpha = 0.05, returnSamples = F)
}
\arguments{
\item{estimates}{List of MC estimates, provided by the estMC function. If
this is a named list (recommended), the function will use these names in
labeling the differences.}

\item{nSamples}{If NULL (default), compares all MC samples from each pair of
MC estimates.  Otherwise an integer, number of samples (with replacement)
to draw from each pair.}

\item{alpha}{Level for confidence/credible intervals provided.}

\item{returnSamples}{Should the function return all the sampled differences?
Defaults to FALSE to reduce storage requirements. Change to TRUE to
compute your own summary statistics.}
}
\description{
Pairwise differences between two or more independent MC estimates
}
