% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estConnectivity.R
\name{diffMC}
\alias{diffMC}
\title{Pairwise differences between two or more independent MC estimates}
\usage{
diffMC(estimates, nSamples = 1e+05, alpha = 0.05, returnSamples = F)
}
\arguments{
\item{estimates}{List of at leat two MC estimates, provided by the estMC
function. If this is a named list (recommended), the function will use
these names in labeling the differences.}

\item{nSamples}{A positive integer, number of samples (with replacement)
to draw from each pair of MC estimates (default 100000).  If set to NULL,
compares all MC samples from each pair.}

\item{alpha}{Level for confidence/credible intervals provided.}

\item{returnSamples}{Should the function return all the sampled differences?
Defaults to FALSE to reduce storage requirements. Change to TRUE to
compute your own summary statistics.}
}
\description{
Pairwise differences between two or more independent MC estimates
}
