% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estConnectivity.R
\name{estMantel}
\alias{estMantel}
\title{Estimate Mantel correlation (rM) from geolocator and/or GPS data.}
\usage{
estMantel(targetPoints, originPoints, isGL, geoBias = NULL, geoVCov = NULL,
  targetSites = NULL, nBoot = 1000, nSim = 1000, verbose = 0,
  alpha = 0.05,
  resampleProjection = "+proj=eqc +lat_ts=0 +lat_0=0 +lon_0=0 +x_0=0 +y_0=0 +a=6371007 +b=6371007 +units=m +no_defs")
}
\arguments{
\item{targetPoints}{A \code{SpatialPoints} object, with length number of
animals tracked.  Each point indicates the point estimate location in
the non-release season.}

\item{originPoints}{A \code{SpatialPoints} object, with length number of
animals tracked.  Each point indicates the release location of an animal.}

\item{isGL}{Indicates whether or which animals were tracked with geolocators.
Should be either single TRUE or FALSE value, or vector with length of
number of animals tracked, with TRUE for animals in
\code{targetPoints} with geolocators and FALSE for animals with GPS.}

\item{geoBias}{For GL data, vector of length 2 indicating expected bias
in longitude and latitude of \code{targetPoints}, in
\code{resampleProjection} units (default meters).}

\item{geoVCov}{For GL data, 2x2 matrix with expected variance/covariance
in longitude and latitude of \code{targetPoints}, in
\code{resampleProjection} units (default meters).}

\item{targetSites}{A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame}
object indicating valid target location(s).  Not needed unless you want
to mask out certain areas (e.g. water).}

\item{nBoot}{Number of bootstrap runs. Animals are sampled with replacement for each,
to estimate sampling uncertainty.}

\item{nSim}{Tuning parameter for GL data. Affects only the speed; 1000 seems
to work well with our data.  Should be integer > 0.}

\item{verbose}{0 (default) to 3. 0 prints no output during run. 1 prints
a line every 100 bootstraps.  2 prints a line every bootstrap.
3 also prints the number of draws (for tuning nSim for GL data only).}

\item{alpha}{Level for confidence/credible intervals provided.}

\item{resampleProjection}{Projection when sampling from geolocator bias/error.
   Default Equidistant Conic = "+proj=eqc +lat_ts=0 +lat_0=0 +lon_0=0 +x_0=0 +y_0=0 +a=6371007 +b=6371007 +units=m +no_defs"}
}
\value{
\code{estMantel} returns a list with elements:
\describe{
  \item{\code{sampleCorr}}{\code{nBoot} sampled values for Mantel
     correlation. Provided to allow the user to compute own summary
     statistics.}
  \item{\code{pointCorr}}{Simple point estimate of Mantel correlation,
     using \code{originPoints} and \code{targetPoints}.}
  \item{\code{meanCorr, medianCorr, seCorr, simpleCICorr, bcCICorr}}{Summary
     statistics for Mantel correlation bootstraps.}
}
}
\description{
Resampling of uncertainty for rM from SpatialPoints geolocators and/or GPS
data.
}
\examples{
rM1 <- estMantel(isGL=OVENdata$isGL, # Logical vector for light-level geolocator (TRUE) or GPS (F)
                 geoBias = OVENdata$geo.bias, # Light-level geolocator location bias
                 geoVCov = OVENdata$geo.vcov, # Light-level geolocator co-variance matrix
                 targetSites = OVENdata$targetSites, # Non-breeding target sites
                 originPoints = OVENdata$originPoints, # Capture Locations
                 targetPoints = OVENdata$targetPoints, # Non-breeding Locations derived from devices
                 verbose = 1,   # output options
                 nBoot = 100, # This is set low for example
                 resampleProjection = raster::projection(OVENdata$targetSites))
str(rM1)
}
\seealso{
\code{\link{estMC}}
}

