% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estConnectivity.R
\name{estMCCmrAbund}
\alias{estMCCmrAbund}
\title{Estimate MC from abundance and transition probability estimates.}
\usage{
estMCCmrAbund(originRelAbund, psi, originDist, targetDist, originSites = NULL,
  targetSites = NULL, nSamples = 1000, row0 = 0, verbose = 0,
  alpha = 0.05)
}
\arguments{
\item{originRelAbund}{Relative abundances at B origin sites. Either a
numeric vector of length B that sums to 1 or an mcmc object with
\code{nSamples} rows  and columns including 'relN[1]' through 'relN[B]'.}

\item{psi}{Transition probabilities between B origin and W target sites.
Either a matrix with B rows and W columns where rows sum to 1 or a MARK
object with estimates of transition probabilities.}

\item{originDist}{Distances between the B origin sites.  Symmetric B by B
matrix.}

\item{targetDist}{Distances between the W target sites.  Symmetric W by W
matrix.}

\item{originSites}{If \code{psi} is a MARK object, this must be a numeric
vector indicating which sites are origin.}

\item{targetSites}{If \code{psi} is a MARK object, this must be a numeric
vector indicating which sites are target.}

\item{nSamples}{Number of times to resample \code{psi} and/or
\code{originRelAbund}.}

\item{row0}{If \code{originRelAbund} is an mcmc object, this can be set
to 0 (default) or any greater integer to specify where to stop ignoring
samples ("burn-in").}

\item{verbose}{0 (default), 1, or 2. 0 prints no output during run. 1 prints
a line every 100 samples.  2 prints a line every sample.}

\item{alpha}{Level for confidence intervals provided.}
}
\value{
\code{estMCCmrAbund} returns a list with elements:
\describe{
  \item{\code{sampleMC}}{\code{nSamples} sampled values for MC. Provided
     to allow the user to compute own summary statistics.}
  \item{\code{pointMC}}{Simple point estimate of MC, using the point
     estimates of \code{psi} and \code{originRelAbund}.}
  \item{\code{meanMC, medianMC}}{Mean and median of \code{sampleMC}.
     Estimates of MC incorporating parametric uncertainty.}
  \item{\code{seMC}}{Standard error of MC, estimated from SD of
     \code{sampleMC}.}
  \item{\code{simpleCI}}{\code{1 - alpha} confidence interval for MC,
     estimated as \code{alpha/2} and \code{1 - alpha/2} quantiles of
     \code{sampleMC}.}
  \item{\code{bcCI}}{Bias-corrected \code{1 - alpha} confidence interval
     for MC.  Preferable to \code{simpleCI} when \code{pointMC} is the
     best estimate of MC. \code{simpleCI} is preferred when
     \code{medianMC} is a better estimator. When \code{pointMC==medianMC},
     these should be equivalent.}
  \item{\code{hpdCI}}{\code{1 - alpha} credible interval for MC,
     estimated using the highest posterior density (HPD) method.}
}
}
\description{
Resampling of uncertainty for MC from RMark psi matrix estimates and/or JAGS
relative abundance MCMC samples.
}

