% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estConnectivity.R
\name{idSites}
\alias{idSites}
\title{Find optimal sites (regions, populations)}
\usage{
idSites(
  originGrid = NULL,
  targetGrid = NULL,
  originBlocks = NULL,
  targetBlocks = NULL,
  originRange = NULL,
  targetRange = NULL,
  originPoints = NULL,
  targetPoints = NULL,
  nSamples = 1000,
  isGL = FALSE,
  isTelemetry = FALSE,
  isRaster = FALSE,
  captured = "origin",
  geoBias = NULL,
  geoVCov = NULL,
  geoBiasOrigin = geoBias,
  geoVCovOrigin = geoVCov,
  targetRaster = NULL,
  originRaster = NULL,
  verbose = 0,
  alpha = 0.05,
  resampleProjection = "ESRI:102010",
  nSim = ifelse(any(isRaster & isGL), 5000, ifelse(any(isGL), 1000, ifelse(any(isRaster),
    10, 1))),
  maxTries = 300,
  nBurnin = 5000,
  nChains = 3,
  nThin = 1,
  dataOverlapSetting = c("dummy", "none", "named"),
  originRelAbund = NULL,
  targetRelAbund = NULL,
  algorithm = c("maxMC", "cluster"),
  method = c("bootstrap", "MCMC"),
  originSites = NULL,
  targetSites = NULL,
  originNames = NULL,
  targetNames = NULL,
  returnAllInput = TRUE
)
}
\arguments{
\item{returnAllInput}{}
}
\value{
list
}
\description{
Find optimal sites (regions, populations)
}
