% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoAssign.R
\name{isoAssign}
\alias{isoAssign}
\title{IsoAssign
The \code{IsoAssign} function does ....}
\usage{
isoAssign(isovalues, isoSTD, intercept, slope, oddsRatio = FALSE,
  odds = NULL, SingleCellAssign = FALSE, nSim = NULL, dataFrame = FALSE,
  sppShapefile = NULL, assignExtent = c(-179, -60, 15, 89),
  return = "probability", element = "Hydrogen", surface = FALSE,
  period = "Annual")
}
\arguments{
\item{isovalues}{vector of tissue isotope values}

\item{isoSTD}{standard deviation from calibration}

\item{intercept}{intercept value from calibration}

\item{slope}{value from calibration}

\item{oddsRatio}{logical to convert probability into odds ratio}

\item{odds}{if \code{oddsRatio} the odds ratio to use to set likely and unlikely locations}

\item{nSim}{integer specifying how many random samples to draw from a multinomial distribution}

\item{dataFrame}{logical defining whether to return results as a data.frame. If FALSE (default) returns a
raster layer.}

\item{sppShapefile}{SpatialPolygon layer defining species range. Assignments are restricted to these
areas.}

\item{assignExtent}{definition for the extent of the assignment. Can be used in place of \code{sppShapefile} to
limit assignment. Input should follow \code{c(xmin,xmax,ymin,ymax)} in degrees longitude and latitude.}

\item{return}{which assignment map to return. 'probability' returns assignment probability.
'odds' returns a single map for each isovalue as likely or unlikely assignment. 'population'
the population level assignment (sum of all 'odds').}

\item{element}{The elemental isotope of interest. Currently the only
elements that are implemented are 'Hydrogen' (default) and 'Oxygen'}

\item{surface}{if "TRUE" returns surface water values. Defaults is 'FALSE'
which returns the isotopes ratio found in precipitation.}

\item{period}{The time period of interest. If 'Annual' returns a raster
of mean annual values in precipitation for the \code{element}. If
'GrowingSeason' returns growing season values in precipitation for
 \code{element} of interest.}

\item{singleCellAssign}{if TRUE generates \code{nSim} single location assignments using a multinomial
distribution using the assignment probability.}
}
\value{
raster stack, raster layer or data.frame if \code{dataFrame = TRUE} of assignment probabilites
}
\description{
IsoAssign
The \code{IsoAssign} function does ....
}
