% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrinsicIsotopes.R
\name{isoAssign}
\alias{isoAssign}
\title{isoAssign
The \code{isoAssign} function generates origin assignments using stable-hydrogen isotopes in tissue. The function generates
a probability surface of origin assignment from a vector of stable-isotope values for each animal/sample of interest.
Probabilistic assignments are constructed by first converting observed stable-isotope ratios (isoscape) in either precipitation or surface
waters into a 'tissuescape' using a user-provided intercept, slope and standard deviation. See \href{http://journals.plos.org.mutex.gmu.edu/plosone/article?id=10.1371/journal.pone.0035137}{Hobson et. al. 2012}}
\usage{
isoAssign(isovalues, isoSTD, intercept, slope, odds = 0.33,
  restrict2Likely = TRUE, nSamples = NULL, sppShapefile = NULL,
  assignExtent = c(-179, -60, 15, 89), element = "Hydrogen",
  surface = FALSE, period = "Annual", seed = NULL)
}
\arguments{
\item{isovalues}{vector of tissue isotope values}

\item{isoSTD}{standard deviation from calibration}

\item{intercept}{intercept value from calibration}

\item{slope}{value from calibration}

\item{odds}{odds ratio to use to set likely and unlikely locations defaults to 0.33}

\item{restrict2Likely}{if \code{TRUE} restricts locations to fall within the 'likely' assignment
locations.}

\item{nSamples}{integer specifying how many random samples to draw from a multinomial distribution.}

\item{sppShapefile}{SpatialPolygon layer defining species range. Assignments are restricted to these
areas.}

\item{assignExtent}{definition for the extent of the assignment. Can be used in place of \code{sppShapefile} to
limit assignment. Input should follow \code{c(xmin,xmax,ymin,ymax)} in degrees longitude and latitude.}

\item{element}{The elemental isotope of interest. Currently the only
elements that are implemented are 'Hydrogen' (default) and 'Oxygen'}

\item{surface}{if "TRUE" returns surface water values. Defaults is 'FALSE'
which returns the isotopes ratio found in precipitation.}

\item{period}{The time period of interest. If 'Annual' returns a raster
of mean annual values in precipitation for the \code{element}. If
'GrowingSeason' returns growing season values in precipitation for
 \code{element} of interest.}

\item{seed}{numeric value fed to \code{set.seed} for random number generation. Default = NULL.}
}
\value{
returns an \code{isoAssign} object containing the following:
    1. \code{probassign} raster stack of individual probabilistic assignments,
    2. \code{oddsassign} raster stack that includes likely vs unlikely origin for each animal,
    3. \code{popassign} a raster for population level assignment (sum of # 2),
    4. \code{probDF} data.frame of individual probability surfaces,
    5. \code{oddsDF} data.frame of likely vs unlikley surfaces,
    6. \code{popDF} data.frame of population level assignment &
    7. \code{SingeCell} array of coordinates (longitude,latitude) for single cell assignment
}
\description{
isoAssign
The \code{isoAssign} function generates origin assignments using stable-hydrogen isotopes in tissue. The function generates
a probability surface of origin assignment from a vector of stable-isotope values for each animal/sample of interest.
Probabilistic assignments are constructed by first converting observed stable-isotope ratios (isoscape) in either precipitation or surface
waters into a 'tissuescape' using a user-provided intercept, slope and standard deviation. See \href{http://journals.plos.org.mutex.gmu.edu/plosone/article?id=10.1371/journal.pone.0035137}{Hobson et. al. 2012}
}
