% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigConnectivity.R
\docType{package}
\name{MigConnectivity}
\alias{MigConnectivity}
\title{MigConnectivity: A package for quantifying migratory connectivity pattern and
strength for migratory animals}
\description{
The MigConnectivity package allows the user to estimate transition
probabilities for migratory animals between any two phases of the annual
cycle, using a variety of different data types with the function
\code{\link{estTransition}}. Also calculates or estimates the strength of
migratory connectivity (MC), a standardized metric to quantify the extent to
which populations co-occur between two phases of the annual cycle. MC is
independent of data type and accounts for the relative abundance of
populations distributed across a seasonal range.  MC is applicable between
any two phases of the annual cycle that are temporally distinct, including
migratory phases. MC can be defined for entire seasonal ranges or for
subsets of those ranges, as long as those subsets of ranges can be further
divided into regional populations. The package includes functions to
estimate MC and the more traditional metric of migratory connectivity
strength (Mantel correlation; rM) incorporating uncertainty from multiple
sources of sampling error. Description of the MC metric can be found in Cohen
et al. (2018).
}
