% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConnectivity.R
\name{optimSites}
\alias{optimSites}
\title{Optimal site arrangement on origin and/or target side}
\usage{
optimSites(
  originSamples,
  targetSamples,
  originBlocks,
  targetBlocks,
  algorithm = "maxMC",
  originRelAbund = NULL,
  originFixed = FALSE,
  targetFixed = FALSE,
  originTies = NULL,
  targetTies = NULL
)
}
\arguments{
\item{targetTies}{}
}
\value{
list
}
\description{
Optimal site arrangement on origin and/or target side
}
