% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputConnectivity.R
\name{plot.intrinsicAssign}
\alias{plot.intrinsicAssign}
\title{Basic plot function for the different isoAssign outputs}
\usage{
\method{plot}{intrinsicAssign}(x, map, ...)
}
\arguments{
\item{x}{an isoAssign object}

\item{map}{which \code{isoAssign} output to plot either 'probability', 'population' or 'odds'}

\item{...}{additional arguments passed to plot function}
}
\value{
A basic plot of the isotope assignments. If \code{map = 'population'} returns a single map.
If \code{map = 'probability' or map = 'odds'} a map for each individual is returned. User is asked for input before each individual is drawn.
}
\description{
Basic plot function for the different isoAssign outputs
}
\examples{
if (interactive()) {
  OVENdist <- terra::vect("data-raw/Spatial_Layers/OVENdist.shp")
  OVENdist <- OVENdist[OVENdist$ORIGIN==2,] # only breeding
  terra::crs(OVENdist) <-
    "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

  OVENvals <- read.csv("data-raw/deltaDvalues.csv")

  b <- isoAssign(isovalues = OVENvals[,2],
                 isoSTD = 12,
                 intercept = -10,
                 slope = 0.8,
                 odds = NULL,
                 restrict2Likely = TRUE,
                 nSamples = 1000,
                 sppShapefile = OVENdist,
                 assignExtent = c(-179,-60,15,89),
                 element = "Hydrogen",
                 surface = FALSE,
                 period = "Annual")

  plot(b, map = "population")
}

}
\seealso{
{\code{isoAssign}}
}
