% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simConnectivity.R
\name{simGeneticData}
\alias{simGeneticData}
\title{Simulate animal genoscape data for estimating migratory connectivity}
\usage{
simGeneticData(
  genPops,
  psi,
  originRelAbund,
  sampleSize,
  originSites = NULL,
  targetSites = NULL,
  captured = "target",
  requireEveryOrigin = FALSE,
  verbose = 0
)
}
\arguments{
\item{requireEveryOrigin}{}
}
\value{

}
\description{
Simulate animal genoscape data for estimating migratory connectivity
}
