% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrinsicIsotopes.R
\name{weightAssign}
\alias{weightAssign}
\title{Calculate Weights for Isotope Assignments
weightAssign}
\usage{
weightAssign(knownLocs, isovalues, isoSTD, intercept, slope, odds = 0.67,
  relativeAbundance, weightRange = c(-1, 1), sppShapefile = NULL,
  assignExtent = c(-179, -60, 15, 89), element = "Hydrogen",
  surface = FALSE, period = "Annual")
}
\arguments{
\item{isovalues}{vector of tissue isotope values from known locations}

\item{isoSTD}{standard deviation from calibration}

\item{intercept}{intercept value from calibration}

\item{slope}{value from calibration}

\item{odds}{odds ratio to use to set likely and unlikely locations defaults to 0.67}

\item{relativeAbundance}{raster layer of relative abundance that sums to 1.}

\item{weightRange}{vector of length 2 within minimum and maximum values to weight isotope and relative abundance.
Default = c(-1,1)}

\item{sppShapefile}{SpatialPolygon layer defining species range. Assignments are restricted to these
areas.}

\item{assignExtent}{definition for the extent of the assignment. Can be used in place of \code{sppShapefile} to
limit assignment. Input should follow \code{c(xmin,xmax,ymin,ymax)} in degrees longitude and latitude.}

\item{element}{The elemental isotope of interest. Currently the only
elements that are implemented are 'Hydrogen' (default) and 'Oxygen'}

\item{surface}{if "TRUE" returns surface water values. Defaults is 'FALSE'
which returns the isotopes ratio found in precipitation.}

\item{period}{The time period of interest. If 'Annual' returns a raster
of mean annual values in precipitation for the \code{element}. If
'GrowingSeason' returns growing season values in precipitation for
 \code{element} of interest.}

\item{knownlocs}{matrix of capture locations of the same length as \code{isovalues}}
}
\value{
returns an \code{weightAssign} object containing the following:
    1. \code{top} data.frame with the optimal weightings,
    2. \code{frontier} data.frame with values that fall along the Pareto frontier,
    3. \code{performance} data.frame with error rate and assignment area for each weight combination
}
\description{
The primary purpose of this function is to determine whether weighting likelihood based isotope assignments
and prior information, such as relative abundance can improve the model performance compared to the
isotope-only model. To do this, we raise the likelihood and prior values to powers from 10^−1
to 10 and measure model performance using the assignment error rate and assignment area. Weights < 1 flatten
the likelihood/prior distributions (giving relatively more weight to smaller values) and weights > 1
sharpen the distributions (giving relatively less weight to smaller values. The \code{weightAssign} function
generates origin assignments using stable-hydrogen isotopes in tissue. If first generates
a probability surface of origin assignment from a vector of stable-isotope values for each animal/sample
captured at a known location. Probabilistic assignments are constructed by first converting observed
stable-isotope ratios (isoscape) in either precipitation or surface waters into a 'tissuescape' using
a user-provided intercept, slope and standard deviation. See \href{http://journals.plos.org.mutex.gmu.edu/plosone/article?id=10.1371/journal.pone.0035137}{Hobson et. al. 2012}
}
\details{
See \href{https://onlinelibrary.wiley.com/doi/10.1002/ece3.2605}{Rushing et al. 2017} for more information.
}
